+++ src/socketCS.cpp
@@ -127,7 +127,7 @@ std::string Socket::ReceiveLine() {
 
 void Socket::SendLine(std::string s) {
   s += '\n';
-  send(s_,s.c_str(),s.length(),0);
+  send(s_,s.c_str(),(int)s.length(),0);
 }
 
 void Socket::SendBytes(const char *buf, int len) {
